Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@8981 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/states_screens/kart_selection.cpp
@@ -343,7 +343,7 @@ public:
         const KartModel &kart_model = props->getMasterKartModel();
         
         m_model_view->addModel( kart_model.getModel(), Vec3(0,0,0), Vec3(35.0f, 35.0f, 35.0f),
-                                (float)kart_model.getBaseFrame() );
+                                kart_model.getBaseFrame() );
         m_model_view->addModel( kart_model.getWheelModel(0), 
                                 kart_model.getWheelGraphicsPosition(0) );
         m_model_view->addModel( kart_model.getWheelModel(1),
